Update LAB_07_Implementing_and_configuring_network_infrastructure_services_in_Windows_Server.md
Guys, whoever created this module 07 description - both standard & extended ak version - would it be so kind to test what he/she described on a real life environment? There is a statement "In the Tools list, browse to the DHCP tool and install the DHCP PowerShell tools." after the Roles & Features DHCP Installation. Guys the DHCP Roles & Features has nothing to do with the "Tools" in WAC. It is a DHCP role on a server. DHCP in WAC Tools is an extension. Installed in a completely different way  in a completely different place - WAC Settings. There is no DHCP PowerShell Tools there. First to see DHCP Tools you have to install a DHCP extension. And then there are scopes not DHCP Tools there. The official MS materials have to be accurate not misleading completely. Would someone be so kind to update that and test as well ?