Update LAB_AK_07_Implementing_and_configuring_network_infrastructure_services_in_Windows Server.md
Guys, really? Who wrote this "On SEA-ADM1, switch back to the Ethernet Status window, and then select Disable.
Back in the Network and Sharing Center window, select Change adapter settings, select Ethernet, and then select Enable this network device." ???
When you do that on a real life server - Disable Network interface to obtain a new IP from the DHCP - you just disconnected yourself from the server and you won't be able to login again. You just lost you access. If the server is in a far remote location where there is no one from IT on site and there is no Out-Of-Band Management network/ILO you just found yourself a back and forth trip there to manually enable the interface. Who wrote that ??? It should be "To release and renew the IP address from the DHCP run from CMD/PowerShell ipconfig /release and next ipconfig /renew.